requirements.txt
pytelegrambotapi
telebot
bs4
pyyaml
pillow
